move more slash commands
