+++ package.json
@@ -38,7 +38,7 @@
 		"jwt-simple": "^0.5.6",
 		"klasa": "https://github.com/gc/klasa.git#99ee76765f6b6172254cbeb363c73f4305d43f36",
 		"klasa-decorators": "^0.0.1",
-		"mahoji": "^0.0.47",
+		"mahoji": "^0.0.48",
 		"math-expression-evaluator": "^1.3.14",
 		"murmurhash": "^2.0.0",
 		"node-cron": "^3.0.0",
